+++ CHANGELOG.md
@@ -20,6 +20,8 @@ some default library
 
 ### 0.14.1
 
+* Replace `UpdatesFilter` and `UpdatesPoller` into another package
+
 ## 0.13.0 Telegram Polls
 
 * Type `PollOption` and `AnonymousPollOption` added